took out some comments
I think the curve is approximating the midpoint too low from the ideal midpoint with the example points used.  This is probably because I'm not breaking the problem into halves, but instead I'm just expanding the total number of points.  
I'm doing this for a class so I'll check with the prof to see if it's satisfactory before I make any more iterations. 
Good night. I just turned 24.